requirements.txt
tinytuya==1.12.10
mss==9.0.1
Pillow==10.4.0
python-dotenv==1.0.1
